backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kiteconnect==5.0.1
fyers-apiv3==3.1.5
websockets==12.0
pydantic==2.5.0
python-dotenv==1.0.0